:recycle: [Refactor] 관리자 관련 폴더 이동 #main
