+++ src/refactoring/components/admin/product/NewProductForm.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { Product } from "../../../types";
+import { Product } from "../../../../types";
 
 interface NewProductFormProps {
   onProductAdd: (product: Product) => void;
