+++ src/refactoring/components/admin/product/ProductCard.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Discount } from "../../../types";
+import { Discount } from "../../../../types";
 
 interface ProductCardProps {
   discount: Discount;
