+++ src/refactoring/components/admin/product/ProductCardList.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Product } from "../../../types";
+import { Product } from "../../../../types";
 import { ProductCard } from "./ProductCard";
 
 interface ProductCardListProps {
