+++ src/refactoring/components/admin/product/ProductHeader.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Product } from "../../../types";
+import { Product } from "../../../../types";
 
 interface ProductHeaderProps {
   product: Product;
